Se cambio el script de la base de datos para aceptar la nueva tabla "jornada"
Se implemento un generador de bloques horarios y un endpoint para consultar la disponibilidad

+++ Producto/backend/src/main/java/com/nodevet/app/dto/AdminDTO.java
@@ -8,7 +8,7 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 @AllArgsConstructor
 public class AdminDTO {
-    private Long idAdmin; // ID del perfil de Admin
+    private Integer idAdmin; // ID del perfil de Admin
     private Integer idUsuario; // ID del Usuario asociado (Integer, como en tu modelo actual)
     private String nombreCompleto;
     private String correoUsr;
